requirements.txt
nltk
pytorch